pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nhl-excite-o-meter"
version = "0.0.0"
requires-python = ">=3.11"
dependencies = [
  "flask",
  "flask-cors",
  "psycopg2",
  "requests",
  "numpy",
  "boto3",
  "pytest-json-ctrf"
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
